removed double() from loss fn

+++ segmentation_code/loss.py
@@ -9,11 +9,9 @@ class BCEDiceLoss(nn.Module):
         truth = target.view(-1)
 
         # BCE loss
-        bce_loss = nn.BCELoss()(pred, truth).double()
+        bce_loss = nn.BCELoss()(pred, truth)
 
         # Dice Loss
-        dice_coef = (2.0 * (pred * truth).double().sum() + 1) / (
-                    pred.double().sum() + truth.double().sum() + 1
-                    )
+        dice_coef = (2.0 * (pred * truth).sum() + 1) / (pred.sum() + truth.sum() + 1)
 
         return bce_loss + (1 - dice_coef)
